Add JA Avail Tracker Project and adjust styling for easier visibility
Added the data for the JA Availability Tracker project to the project.
Need to update the links for it once i have added a repo for it.

Adjusted background color of the Project  cards so they are more
distinguishable from the page.  Chose  linen because light gray was too
dark and linen is darker than the white page but not so dark that it
conflicts with card content.